feat: token generator #35
I created the token generation in the login method, and also created a login model. I didn't send the Json login with user data because I don't need it yet. But if there is such a need, it will be necessary to solve the problem with targeting when creating a json object